1. fix observer structure 2. improve throttler 3. fix unit test of scroll event
