+++ __tests__/web/scroll/scroll.html
@@ -1,8 +1,8 @@
 <head>
   <style>
     #container {
-      width: 100px;
-      height: 100px;
+      width: 500px;
+      height: 500px;
       overflow: auto;
     }
 
@@ -15,8 +15,6 @@
 
 <body>
   <div id="container">
-    <div id="content">
-
-    </div>
+    <div id="content"></div>
   </div>
 </body>
